requirements.txt
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
idna==2.10
multidict==5.1.0
pyTelegramBotAPI==3.7.4
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.1
selenium==3.141.0
soupsieve==2.1
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
